+++ unnamed_engine/src/gui/egui_renderer.rs
@@ -1,3 +1,6 @@
+//! ## Egui Renderer
+//!
+//! Responsible for rendering EGUI into wgpu.
 use egui::epaint::Shadow;
 use egui::{Context, Visuals};
 use egui_wgpu::{Renderer, ScreenDescriptor};
@@ -64,7 +67,7 @@ impl EguiRenderer {
     run_ui: impl FnOnce(&Context),
   ) {
     let raw_input = self.state.take_egui_input(&window);
-    let full_output = self.context.run(raw_input, |ui| {
+    let full_output = self.context.run(raw_input, |_ui| {
       run_ui(&self.context);
     });