+++ unnamed_engine/src/gui/mod.rs
@@ -1,2 +1,3 @@
+//! ## GUI
 pub mod egui_renderer;
 pub mod gui;